The codebase was cleaned and config.json added for one line initializing.

+++ start_sim.sh
@@ -1,13 +1,19 @@
 #!/bin/bash
 
+# Read drone count from config.json
+DRONE_COUNT=$(jq -r '.uav_count' config.json)
+
+if [ -z "$DRONE_COUNT" ]; then
+    echo "Hata: uav_count config.json içinde bulunamadı veya okunamadı."
+    exit 1
+fi
+
 killall -9 gzserver gzclient px4 &> /dev/null
 
 if [ -z "$PX4_PATH" ]; then
     echo "Hata: PX4_PATH değişkeni tanımlı değil. Lütfen ~/.bashrc dosyanızda örneğin : export PX4_PATH=~/PX4-Autopilot şeklinde tanımlayın."
     exit 1
 fi
-
-DRONE_COUNT=5
 
 for (( i=0; i<DRONE_COUNT; i++ ))
 do
